Removed the parsing functions from the class to utilities
