+++ FileServerSubsystem.h
@@ -23,17 +23,4 @@ public:
     ~FileServerSubsystem() override = default;
 
     Poco::UInt16 getPort() const;
-
-
-private:
-    struct NumberAndDimension
-    {
-        std::string_view number;
-        std::string_view dimension;
-    };
-
-    std::size_t getAsBytesSize(const std::string& size_str) const;
-    NumberAndDimension parseDimensionFormat(std::string_view argument) const;
-    std::size_t getSizeDimension(std::string_view dim_part) const;
-    std::size_t getNumber(std::string_view number_part) const;
 };
