+++ utils/utils.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <iostream>
+
+namespace parsing
+{
+    std::size_t getAsBytesSize(const std::string& size_str);
+}
